drm/nouveau/bios/init: bump script offset to 32-bits
No (known) case yet, but other tables have been moving beyond 16-bits,
so we may as well be prepared.

+++ drivers/gpu/drm/nouveau/include/nvkm/subdev/bios/init.h
@@ -5,7 +5,7 @@ FASTVC-BENCH-CORPUS:linux-main-100k-v1-e0bd41dc-8e12-4f1b-978d-a3645ae59da0
 struct nvbios_init {
 	struct nvkm_subdev *subdev;
 	struct nvkm_bios *bios;
-	u16 offset;
+	u32 offset;
 
 	struct dcb_output *outp;
 	union {
@@ -16,8 +16,8 @@ struct nvbios_init {
 	/* internal state used during parsing */
 	u8 execute;
 	u32 nested;
-	u16 repeat;
-	u16 repend;
+	u32 repeat;
+	u32 repend;
 	u32 ramcfg;
 };
